Revises Tor apt mirror URL
The backend recently moved around, so updating the deb package
implementation to reflect the backend changes. Updated the URL from:

    tor-apt.ops.freedom.press

to:

    tor-apt.freedom.press

We also have "tor-apt-test.freedom.press" for QA purposes, but without templates
for the Debian scripts, however, swapping the URLs for QA will need to be manual.
We can circle back and smooth out that process in subsequent PRs—the
priority now is to make sure we have mirror functionality working.