i18n: do not expose unsupported locales
The SUPPORTED_LOCALES list is enforced so the user cannot select a
language that it does not contain. The restriction based on the
SUPPORTED_LOCALES variable is done at setup and used to narrow the
content of the LOCALES variable.

The SUPPORTED_LOCALES content was previously only used to select the
languages displayed via the language menu. This is incorrect because
it would allow a user to add ?l=unsupported to see the interface in
one of the available languages even if it is not in the
SUPPORTED_LOCALES list. This was also true if the browser it configured
to prefer a given locale.